Вынес шаблон для создания БД в отдельный файл
